Move helpers to integration test package part 1 (WIP)
