+++ pkg/sdk/integration_tests/random.go
@@ -5,11 +5,40 @@ import (
 
 	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/sdk"
 	"github.com/brianvoe/gofakeit/v6"
+	"github.com/hashicorp/go-uuid"
+	"github.com/stretchr/testify/require"
 )
+
+func randomUUID(t *testing.T) string {
+	t.Helper()
+	v, err := uuid.GenerateUUID()
+	require.NoError(t, err)
+	return v
+}
+
+func randomComment(t *testing.T) string {
+	t.Helper()
+	return gofakeit.Sentence(10)
+}
+
+func randomBool(t *testing.T) bool {
+	t.Helper()
+	return gofakeit.Bool()
+}
 
 func randomString(t *testing.T) string {
 	t.Helper()
 	return gofakeit.Password(true, true, true, true, false, 28)
+}
+
+func randomStringN(t *testing.T, num int) string {
+	t.Helper()
+	return gofakeit.Password(true, true, true, true, false, num)
+}
+
+func randomAlphanumericN(t *testing.T, num int) string {
+	t.Helper()
+	return gofakeit.Password(true, true, true, false, false, num)
 }
 
 func randomStringRange(t *testing.T, min, max int) string {
@@ -28,9 +57,14 @@ func randomIntRange(t *testing.T, min, max int) int {
 	return gofakeit.IntRange(min, max)
 }
 
-func randomAlphanumericN(t *testing.T, num int) string {
+func randomSchemaObjectIdentifier(t *testing.T) sdk.SchemaObjectIdentifier {
 	t.Helper()
-	return gofakeit.Password(true, true, true, false, false, num)
+	return sdk.NewSchemaObjectIdentifier(randomStringN(t, 12), randomStringN(t, 12), randomStringN(t, 12))
+}
+
+func randomDatabaseObjectIdentifier(t *testing.T) sdk.DatabaseObjectIdentifier {
+	t.Helper()
+	return sdk.NewDatabaseObjectIdentifier(randomStringN(t, 12), randomStringN(t, 12))
 }
 
 func alphanumericDatabaseObjectIdentifier(t *testing.T) sdk.DatabaseObjectIdentifier {
@@ -38,7 +72,7 @@ func alphanumericDatabaseObjectIdentifier(t *testing.T) sdk.DatabaseObjectIdenti
 	return sdk.NewDatabaseObjectIdentifier(randomAlphanumericN(t, 12), randomAlphanumericN(t, 12))
 }
 
-func randomComment(t *testing.T) string {
+func randomAccountObjectIdentifier(t *testing.T) sdk.AccountObjectIdentifier {
 	t.Helper()
-	return gofakeit.Sentence(10)
+	return sdk.NewAccountObjectIdentifier(randomStringN(t, 12))
 }
